Refactor MCP into a module
